+++ CHANGELOG.rst
@@ -1,6 +1,17 @@
 Changelog
 ~~~~~~~~~
 
+1.9.9 - 2023-04-03
+------------------
+
+* Prevent situation in which slug could end in dash or underscore
+* Remove support for end-of-life Python & Django versions
+
+1.9.8 - 2020-07-22
+------------------
+
+Move FieldDoesNotExist import for compatibility with Django 3.1
+
 1.9.7 - 2020-04-14
 ------------------